+++ src/routes/auth/auth.test.ts
@@ -10,7 +10,9 @@ import {
   REDIRECT_URL_ERROR,
   JWT_CLAIMS_NAMESPACE,
   HOST,
-  PORT
+  PORT,
+  ADMIN_SECRET_HEADER,
+  HASURA_GRAPHQL_ADMIN_SECRET
   // ANONYMOUS_USERS_ENABLE
 } from '@shared/config'
 import { generateRandomString, selectAccountByEmail } from '@shared/helpers'
@@ -59,7 +61,57 @@ it('should create an account', async () => {
   const { status } = await request
     .post('/auth/register')
     .send({ email, password, user_data: { name: 'Test name' } })
-  expect(status).toEqual(204)
+  expect(status).toEqual(200)
+})
+
+it('should fail to create account with unallowed role', async () => {
+  const { status } = await request.post('/auth/register').send({
+    email: 'test1@nhost.io',
+    password,
+    user_data: { name: 'Test name' },
+    register_options: {
+      allowed_roles: ['user', 'me', 'super-admin']
+    }
+  })
+  expect(status).toEqual(400)
+})
+
+it('should fail to create accunt with default_role that does not overlap allowed_roles', async () => {
+  const { status } = await request.post('/auth/register').send({
+    email: 'test2@nhost.io',
+    password,
+    user_data: { name: 'Test name' },
+    register_options: {
+      default_role: 'editor',
+      allowed_roles: ['user', 'me']
+    }
+  })
+  expect(status).toEqual(400)
+})
+
+it('should create account with default_role that is in the ALLOWED_USER_ROLES variable', async () => {
+  const { status } = await request.post('/auth/register').send({
+    email: 'test3@nhost.io',
+    password,
+    user_data: { name: 'Test name' },
+    register_options: {
+      default_role: 'editor'
+    }
+  })
+  expect(status).toEqual(200)
+})
+
+it('should register account with default_role and allowed_roles set', async () => {
+  const { status } = await request.post('/auth/register').send({
+    email: 'test4@nhost.io',
+    password,
+    user_data: { name: 'Test name' },
+    register_options: {
+      default_role: 'user',
+      allowed_roles: ['user', 'me']
+    }
+  })
+  expect(status).toEqual(200)
 })
 
 it('should tell the account already exists', async () => {
@@ -125,6 +177,26 @@ it('should sign the user in', async () => {
   expect(body.jwt_expires_in).toBeNumber()
 })
 
+it('should not sign user in with invalid admin secret', async () => {
+  const { status } = await request
+    .post('/auth/login')
+    .set(ADMIN_SECRET_HEADER, 'invalidsecret')
+    .send({ email, password })
+
+  expect(status).toEqual(401)
+})
+
+it('should sign user in with valid admin secret', async () => {
+  const { body, status } = await request
+    .post('/auth/login')
+    .set(ADMIN_SECRET_HEADER, HASURA_GRAPHQL_ADMIN_SECRET as string)
+    .send({ email, password: 'invalidpassword' })
+
+  expect(status).toEqual(200)
+  expect(body.jwt_token).toBeString()
+  expect(body.jwt_expires_in).toBeNumber()
+})
+
 it('should decode a valid custom user claim', async () => {
   const decodedJwt = JWT.decode(jwtToken) as Token
   expect(decodedJwt[JWT_CLAIMS_NAMESPACE]).toBeObject()
@@ -144,7 +216,6 @@ describe('Tests without cookies', () => {
       .post('/auth/login')
       .send({ email, password, cookie: false })
     // Save JWT token to globally scoped varaible.
-    console.log(body)
     jwtToken = body.jwt_token
     expect(status).toEqual(200)
     expect(body.jwt_token).toBeString()